Improve `code_pegs=` by assigning colors for each codes
Assigning `guess_code` store to the respective row doesn't give their
color values. When the board is drawn the number will appear on the
board but not thire color. This become issue to the interaction between
the players and the board.

Store the guess codes by replacing them with their colors fixes this
issue. This makes the board more interactive.